# src/docs/swagger.yaml
basePath: /
definitions:
...
  ResponseError:
    properties:
      error:
        example: something went wrong
        type: string
    type: object
...
  UsersMeLeaderboardPositionsGetHandler.LeaderboardPosition:
    properties:
      language:
        example: english
        type: string
      mode:
        example: time
        type: string
      rank:
        example: 42
        type: integer
      submode:
        example: 15s
        type: string
    type: object
  UsersMeLeaderboardPositionsGetHandler.ResponseBody:
    properties:
      leaderboardPositions:
        items:
          $ref: '#/definitions/UsersMeLeaderboardPositionsGetHandler.LeaderboardPosition'
        type: array
    type: object
  UsersMeStatisticsGetHandler.ResponseBody:
    properties:
      statistics:
        items:
          $ref: '#/definitions/UsersMeStatisticsGetHandler.Stats'
        type: array
    type: object
  UsersMeStatisticsGetHandler.Stats:
    properties:
      accuracy:
        example: 98.7
        format: float64
        type: number
      cpm:
        example: 210.3
        format: float64
        type: number
      duration:
        example: 120000
        format: int64
        type: integer
      id:
        example: 1
        format: uint
        type: integer
      isPunctuation:
        example: true
        type: boolean
      language:
        example: english
        type: string
      mode:
        example: time
        type: string
      playedAt:
        example: "2023-12-25T15:04:05Z"
        format: date-time
        type: string
      submode:
        example: 30s
        type: string
      uncompletedTestsCount:
        example: 0
        type: integer
      uncompletedTestsTotalDuration:
        example: 0
        type: integer
      wpm:
        example: 42.5
        format: float64
        type: number
    type: object
  UsersMeStatisticsPostHandler.RequestBody:
    properties:
      accuracy:
        example: 98.7
        type: number
      cpm:
        example: 210.3
        type: number
      createdAt:
        example: "2025-10-19T19:02:29+03:00"
        type: string
      durationMs:
        example: 60000
        type: integer
      finishedAt:
        example: "2025-10-19T19:02:29+03:00"
        type: string
      isPunctuation:
        example: true
        type: boolean
      language:
        example: english
        type: string
      mode:
        example: time
        type: string
      sign:
        example: "12345"
        type: string
      startedAt:
        example: "2025-10-19T19:02:29+03:00"
        type: string
      submode:
        example: 1m
        type: string
      uid:
        example: "0"
        type: string
      uncompletedTestsCount:
        example: 0
        type: integer
      uncompletedTestsDurationMs:
        example: 0
        type: integer
      wpm:
        example: 42.5
        type: number
    type: object
  UsersMyStatisticsPostHandler.LeaderboardRankChanged:
    properties:
      from:
        example: 1000
        type: integer
      to:
        example: 921
        type: integer
    type: object
  UsersMyStatisticsPostHandler.ResponseBody:
    properties:
      isPersonalBest:
        type: boolean
      leaderboardRankChanged:
        $ref: '#/definitions/UsersMyStatisticsPostHandler.LeaderboardRankChanged'
      wpmShift:
        type: number
    type: object
...
info:
  contact: {}
  title: Blindtyping
  version: "0.1"
paths:
...
  /users/me/leaderboard-positions:
    get:
      consumes:
      - application/json
      description: Получить местоположение текущего пользователя в лидербордах
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/UsersMeLeaderboardPositionsGetHandler.ResponseBody'
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Получить местоположение текущего пользователя в лидербордах
      tags:
      - Leaderboards
  /users/me/statistics:
    delete:
      consumes:
      - application/json
      description: Делает soft delete для всей статистики пользователя
      produces:
      - application/json
      responses:
        "200":
          description: Статистика пользователя удалена
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/ResponseError'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/ResponseError'
        "500":
          description: Внутренняя ошибка сервера (смотреть логи)
          schema:
            $ref: '#/definitions/ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Отчистить всю статистику пользователя
      tags:
      - User Statistics
    get:
      consumes:
      - application/json
      description: Отдает статистику текущего пользователя
      parameters:
      - description: Дата начала периода выборки в RFC3339
        example: "2015-09-15T14:00:12-00:00"
        in: query
        name: dateFrom
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Статистика пользователя
          schema:
            $ref: '#/definitions/UsersMeStatisticsGetHandler.ResponseBody'
        "400":
          description: Неправильно сформирован запрос
          schema:
            $ref: '#/definitions/ResponseError'
        "401":
          description: Пользователь не авторизован в системе
          schema:
            $ref: '#/definitions/ResponseError'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/ResponseError'
        "500":
          description: Серверная ошибка
          schema:
            $ref: '#/definitions/ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Получить статистику текущего пользователя
      tags:
      - User Statistics
    post:
      consumes:
      - application/json
      description: Creates a new statistics record for the authenticated user
      parameters:
      - description: Statistics data to save
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/UsersMeStatisticsPostHandler.RequestBody'
      produces:
      - application/json
      responses:
        "201":
          description: Statistics successfully created
          schema:
            $ref: '#/definitions/UsersMyStatisticsPostHandler.ResponseBody'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ResponseError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Save typing statistics
      tags:
      - User Statistics
...
swagger: "2.0"